requirements.txt
# Core dependencies
streamlit>=1.28.0
agno>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# AI/ML dependencies
groq>=0.4.0
google-generativeai>=0.3.0

# UI dependencies
streamlit-code-editor>=0.1.0

# Utility dependencies
typing-extensions>=4.0.0